refactor: Delete a copy of an existing module
